servitec_manager/requirements.txt
customtkinter
reportlab
matplotlib
tkcalendar
pillow
pyinstaller
pandas
openpyxl
pdfplumber
pywhatkit
twilio
python-dotenv
